+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 
 - Remove deprecated blockhainAdapterUrl field from PublicConfiguration. (#730)
 - Remove deprecated REST endpoints in TaskController. (#731)
+- Move `WorkerModel` from `iexec-common` to `iexec-core-library`. (#735)
 
 ### Dependency Upgrades